Add py.typed and remove some type ignore tags (#101)
This should ensure that the package works properly with mypy when imported. There are still quite a few type: ignores left for numpy types that should be cleaned up. This also adds a reset_params() function to the abstract language model so that we're not accessing implementation details like parameters directly here.