Change tuples to classes (#103)
The beam-related tuples are getting out of hand and are very difficult to use - changing these to dataclasses. Namedtuples might also be an ok option and are likely to be slightly faster when initializing the objects. Hopefully any performance hit won't be noticeable since this is just packaging data in a slightly different way